// src/api/generated/model/webhookCreationDtoSubscribedEventsItem.ts
/**
 * Generated by orval v7.20.0 🍺
 * Do not edit manually.
 * Cosy API
 * Management API for Cosy (Cost Optimised Server Yard).
 * OpenAPI spec version: v1.0
 */

export type WebhookCreationDtoSubscribedEventsItem = typeof WebhookCreationDtoSubscribedEventsItem[keyof typeof WebhookCreationDtoSubscribedEventsItem];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const WebhookCreationDtoSubscribedEventsItem = {
  SERVER_STARTED: 'SERVER_STARTED',
  SERVER_STOPPED: 'SERVER_STOPPED',
  SERVER_FAILED: 'SERVER_FAILED',
} as const;